+++ package.json
@@ -8,12 +8,12 @@
     "@types/react": "^18.0.26",
     "@types/react-dom": "^18.0.10",
     "axios": "^1.2.2",
-    "bootstrap": "^4.2.1",
+    "bootstrap": "^5.3.0",
     "lodash": "^4.17.11",
     "react": "^18.2.0",
     "react-datepicker": "^2.0.0",
     "react-dom": "^18.2.0",
-    "reactstrap": "^6.5.0"
+    "reactstrap": "^9.1.10"
   },
   "scripts": {
     "start": "vite",
