+++ src/App.tsx
@@ -2,6 +2,5 @@ import { Reservations } from './components'
 import React from 'react'
 
 export const App = () => {
-  return <div>Text</div>
-  // return <Reservations />
+  return <Reservations />
 }
