--- src/actions/actionTypes.js
@@ -1,2 +1,0 @@
-export const CLEAR_RESERVATIONS = "CLEAR_RESERVATIONS";
-export const SAVE_RESERVATIONS = "SAVE_RESERVATIONS";
