+++ src/types.ts
@@ -1,0 +1,13 @@
+export type Reservation = {
+  id: string
+}
+
+export const WEEK_DAYS = [
+  'monday',
+  'tuesday',
+  'wednesday',
+  'thursday',
+  'friday',
+  'saturday',
+  'sunday',
+] as const
